backend/requirements.txt
fastapi
uvicorn[standard]
motor
firebase-admin
youtube-transcript-api
httpx
python-dotenv
pydantic
beautifulsoup4
